The start of an XSPEC interface
The code used to create the wrapper routine used by the Sherpa
interface is useful for building an XSPEC interface, so separate
it out into it's own file. This also changes exactly what is
being wrapped: for the Sherpa case we were wrapping the
jetmain function, but a review of the bhjet.sl file shows me
that we really want to combine jetmain and jetinterp. It
would be possible to use the ISIS approach in Sherpa - that is
provide a Python model evaluator that calls both jetmain and
jetinterp - but it's not obvious if that is sensible for XSPEC.
So we wrap a single routine that does both. This makes it
harder to tweak (e.g. decide to change the "base" grid the
model is evaluated on).

Add a "lmodel" interface for the model, to tell XSPEC how to
create the model. This is based on the bhjet.sl file.

There has been no attempt to build an XSPEC model using this code.